Move winow into its own module
