+++ tests/multinode.rs
@@ -16,9 +16,9 @@ use solana::ncp::Ncp;
 use solana::result;
 use solana::service::Service;
 use solana::signature::{Keypair, KeypairUtil, Pubkey};
-use solana::streamer::{default_window, WINDOW_SIZE};
 use solana::thin_client::ThinClient;
 use solana::timing::duration_as_s;
+use solana::window::{default_window, WINDOW_SIZE};
 use std::cmp::max;
 use std::env;
 use std::fs::{copy, create_dir_all, remove_dir_all};
